# api/OpenAPI.json
{
    "info": {
        "title": "modifiability-scenario-creator-api",
        "version": "0.0.2",
        "license": {
            "name": "Apache-2.0"
        },
        "description": "[Specification JSON](/api-spec)\n\nAPI to create and evaluate software modifiability scenarios"
    },
    "paths": {
...
        "/systems/{id}": {
            "get": {
                "summary": "/systems/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "/systems/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "/systems/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "swagger": "2.0"
}